IE11 code updates
Yeah, IE11 is a pain in my ass. So I changed a few code snippets to make the code work with the default settings of IE11. I added a local jQuery directory for local testing and also in case there are some weird security issues where jQeury cannot be downloaded.